Fix tests to now include function fq name
